@@ -1,17 +1,23 @@
+import "bootstrap/dist/css/bootstrap.css";
+
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client';
+import { createStore, applyMiddleware } from "redux";
+import { Provider } from "react-redux";
+import promiseMiddleware from 'redux-promise';
+
 import './index.css';
 import App from './App';
-import reportWebVitals from './reportWebVitals';
+import rootReducer from "./reducers";
 
-ReactDOM.render(
+const middlewareEnhancer = applyMiddleware(promiseMiddleware)
+const store = createStore(rootReducer, middlewareEnhancer)
+
+const root = ReactDOM.createRoot(document.getElementById('root'));
+root.render(
   <React.StrictMode>
-    <App />
-  </React.StrictMode>,
-  document.getElementById('root')
+    <Provider store={store}>
+      <App />
+    </Provider>
+  </React.StrictMode>
 );
-
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals();
